+++ core/bots/src/main/java/cern/ais/gridwars/MovingBot.java
@@ -20,7 +20,7 @@ public class MovingBot implements PlayerBot {
 		for (int my = 0; my < GameConstants.UNIVERSE_SIZE; my++)
 			for (int mx = 0; mx < GameConstants.UNIVERSE_SIZE; mx++) {
 				if (universeView.belongsToMe(mx, my)) {
-					Long population = universeView.getPopulation(mx, my);
+					int population = universeView.getPopulation(mx, my);
 					if (population > 1) {
 						movementCommands.add(new MovementCommand(universeView.getCoordinates(mx, my), direction, population / 2));
 					}
